add: hidden element depend on permission

+++ app/Http/Controllers/Backend/PermissionController.php
@@ -1,12 +1,14 @@
 <?php
 
-namespace App\Http\Controllers;
+namespace App\Http\Controllers\Backend;
 
 use App\Http\Requests\StorePermissionRequest;
 use App\Http\Requests\UpdatePermissionRequest;
 use App\Models\Permission;
 use App\Services\PermissionService;
 use Illuminate\Http\Request;
+use App\Http\Controllers\Controller;
+use Illuminate\Support\Facades\Gate;
 
 class PermissionController extends Controller
 {
@@ -20,6 +22,7 @@ class PermissionController extends Controller
 
     public function index(Request $request)
     {
+        Gate::authorize('modules', 'permission.index');
         $permissions = $this->permissionService->paginate($request);
         return view('Backend.permission.index', [
             'permissions' => $permissions
@@ -28,11 +31,14 @@ class PermissionController extends Controller
 
     public function create()
     {
+        Gate::authorize('modules', 'permission.create');
+
         return view('Backend.permission.create');
     }
 
     public function store(StorePermissionRequest $request)
     {
+        Gate::authorize('modules', 'permission.create');
         if ($this->permissionService->create($request)) {
             return redirect()->route('permission.index')->with('success', __('alert.addSuccess', ['attribute'=> __('dashboard.permission')]));
         }
@@ -41,6 +47,7 @@ class PermissionController extends Controller
 
     public function edit(Permission $permission)
     {
+        Gate::authorize('modules', 'permission.update');
         return view('backend.permission.create', [
             'permission' => $permission
         ]);
@@ -48,7 +55,7 @@ class PermissionController extends Controller
 
     public function update($id, UpdatePermissionRequest $request)
     {
-
+        Gate::authorize('modules', 'permission.update');
         if ($this->permissionService->update($id, $request)) {
             return redirect()->route('permission.index')->with('success', __('alert.updateSuccess', ['attribute'=> __('dashboard.permission')]));
         }
@@ -58,6 +65,7 @@ class PermissionController extends Controller
 
     public function delete(Permission $permission)
     {
+        Gate::authorize('modules', 'permission.delete');
         return view('backend.permission.delete', [
             'permission' => $permission
         ]);
@@ -65,6 +73,7 @@ class PermissionController extends Controller
 
     public function destroy($id)
     {
+        Gate::authorize('modules', 'permission.delete');
         if ($this->permissionService->destroy($id)) {
             return redirect()->route('permission.index')->with('success', __('alert.deleteSuccess', ['attribute'=> __('dashboard.permission')]));
         }
